sensor-simulations/requirements.txt
azure.iot.device